Refactor EventQueue Tests
