+++ application/Common/SharedTest/Events/Queue/Strategy/SingleConsumer/UnboundedSingleConsumerChannelStrategyTest.cs
@@ -1,22 +1,15 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MORR.Shared.Events.Queue.Strategy.SingleConsumer;
-using SharedTest.TestHelpers.EventQueueStrategy;
 
 namespace SharedTest.Events.Queue.Strategy.SingleConsumer
 {
     [TestClass]
-    public class UnboundedSingleConsumerChannelStrategyTest: EventQueueStorageStrategyTest<UnboundedSingleConsumerChannelStrategy<TestEvent>>
+    public class UnboundedSingleConsumerChannelStrategyTest: SingleConsumerChannelStrategyTest<UnboundedSingleConsumerChannelStrategy<TestEvent>>
     {
         [TestInitialize]
         public void BeforeTest()
         {
             Strategy = new UnboundedSingleConsumerChannelStrategy<TestEvent>();
         }
-
-        [TestMethod]
-        public void TestUnboundedSingleConsumer_Enqueue() => EventQueueStorageStrategyTestClass.Assert_EnqueueSingleProducerUnbounded(Strategy);
-
-        [TestMethod]
-        public void TestUnboundedSingleConsumer_IsOccupied() => SingleConsumerChannelStrategyTestClass.Assert_IsOccupied(Strategy);
     }
 }
